etl/requirements.txt
python-dotenv==1.0.0
pydantic==1.10.7
psycopg2==2.9.5
backoff==2.2.1